Commenting some code, renaming and setting integration tests
