+++ src/main/java/com/github/jhonnymertz/wkhtmltopdf/wrapper/exceptions/WkhtmltopdfConfigurationException.java
@@ -1,0 +1,13 @@
+package com.github.jhonnymertz.wkhtmltopdf.wrapper.exceptions;
+
+
+/**
+ * Exception to describe and track wrapper configuration errors
+ */
+public class WkhtmltopdfConfigurationException extends RuntimeException  {
+
+    public WkhtmltopdfConfigurationException(String s) {
+        super(s);
+    }
+
+}
